+++ Source/World/World_Constants.h
@@ -3,10 +3,10 @@
 
 namespace World
 {
-    constexpr static int    CHUNK_SIZE      = 20,    //Width/ depth of a chunk, height is infinite
-                            WATER_LEVEL     = 90,
-                            BEACH_LEVEL     = WATER_LEVEL + 4,
-                            SNOW_LEVEL      = 260;
+    constexpr static uint16_t   CHUNK_SIZE      = 20,    //Width/ depth of a chunk, height is infinite
+                                WATER_LEVEL     = 90,
+                                BEACH_LEVEL     = WATER_LEVEL + 4,
+                                SNOW_LEVEL      = 260;
 
 }
